// src/api/openapi/model/offerLibrary.ts
/**
 * FastAPI
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { OfferLibraryEntry } from './offerLibraryEntry';


export interface OfferLibrary { 
    description: string;
    entries: Array<OfferLibraryEntry>;
    id: number;
    isManualList: boolean;
    name: string;
}